Move common utils to utils::test_utils
